(TK-384) Docs/test update for watched dir deletion
This commit updates the protocol documentation to more correctly state
the behavior when a watched directory is deleted - namely, it is
unspecified by the API, and platform-dependent in the current
implementation.

As a result, the test case for this is deleted, since the behavior of
the current implementation is so different between OSX and Linux that
it can't be tested in a sane, platform-agnostic way.