(TK-391) Handle OVERFLOW events
Previously we made no distinction for incoming overflow events.

OVERFLOW events do not have a specifc context associated with their
change - only the watchable that the WatchKey was registered with.
And, an event that lacked a context would cause an NPE when
processing the event.

This patch checks to see if the event in question is an OVERFLOW event
and special cases its reporting to return as much information as
possible, without causing an NPE.