(TK-391) Improve overflow-test
Previously, we did not test that the overflow event was handled in a
particular but tested that after doing what we believed would cause an
overflow the watcher correctly recovered. This was necessary due to the
way the patch was implemented in a previous iteration. With an update to
the implementation we can now directly test that an overflow event was
created in a particular way.