(TK-391) Add unit test for overflow event handling
Previously we tested overflow handling at the :integration level. This
has the unfortunate quality of requiring us to cause an error state
within Java/inotify. Up until this point we have been able to do this
easily without major repercussions on the systems that run the test.

That is not true in CI however, where we have a version of the JDK that
will not easily throw overflow events. In fact going to the lengths
required to cause an overflow event on these hosts is just as likely to
crash the JVM or fill up the disk.

This patch adds a unit test to prove our handling of overflow events is
satisfactory, and refactors the production code to make it unit
testable.