(maint) Wrap all event processing with shutdown-fn
Previously, we wrapped the consumer supplied callbacks in registered
shutdown function.

This means that any issue with filesystem watcher code itself would
silently fail in the background, neither logging the issue nor
terminating.

This moves the registered shutdown function to wrap the entire
processing of events so that unforseen issues in the field will be
properly logged and force a clean shutdown.