(TK-389) Move registering new directories closer to discovering them
Previously we registered new directories to watch after we had done all
other processing for them.

Soon we will be adding debouncing, and wiring non-trivial callbacks into
our service. As this happens we will lose the create events from nested
directories.

This moves the watch-new-directories! function to the `retrieve-events!`
function so there is less of a likelihood that events will be missed.