(TK-389) Debounce events per watcher
Previously we reported each event as soon as it occurred as fast as we
could.

This could cause undue load on consumer.

This patch updates our event retreival function to, once an event has
taken place poll again to see if there are waiting watch-keys in the
queue. If there are no waiting watch-keys the poll will block for
`window-min` (100 milliseconds) waiting for events. If still no
additional events occur processing will continue as normal. If an event
is already waiting or occurs during the `window-min` that event will be
added to the list of events we will report. If we have not hit our
`window-max` at that point will repeat our polling procedure.