(TK-389) Consolidate Event representations
Previously passed either a native Java array of WatchEvents, a vector of
clojure maps representing Events, and converted for debugging the native
Java events into a different clojure map format. Calls to the various
formatting happened as needed throughout the retrieval and processing
code causing WatchEvents to be converted several times to clojure
Events.

This patch consolidates our representation of events into a more
comprehensive clojure map that can be created immediately after we receive
them, and can be consumed by all other methods in the core namespace.

To maintain the same logging representations and interface with our
consumers we create several `format-for-*` functions that select
pertinent keys for the representation in question.