(TK-389) Fix file conversion for Event paths
We cannot call `me.raynes.fs/file` on the relative path provided by
`WatchEvent#context()` without it being resolved to an absolute path
relative to the CWD, which is incorrect behavior.

The desired behavior is for the changed-path to remain a relative path,
which is the behavior of calling `.toFile` on the the path, rather than
using `fs/file`.

This path changes our calls from `fs/file` to `.toFile` to preserve
relativeness. Note, this requires moving the calls into conditionals as
changed-path and full-path are nullable and while we can pass `nil`
safely into `fs/file` but we cannot call `.toFile` on nil.