(TK-430) Migrate to clj-parent 0.3.3 and bump i18n to 0.6.0
This commit migrates tk-filesystem-watcher to using clj-parent version 0.3.3
for its dependencies.  This commit also adds i18n as a plugin dependency, to
allow for message POT files to be regenerated as needed in CI.  This
commit also upgrades the lein version to 2.7.1 in the .travis.yml
file since 2.7.1 is the minimum version required for compatibility
with lein/clj-parent.