(TK-448) Add first test that exercises non-recursive functionality
As expected this fails.

Add new test utilities that allow us to check if a specific event - and only
that event - have been registered by the watch service. We do this by
immediately doing our comparison check on the first event that is registered.

We are in the odd position of needing to test that something -doesn't happen-,
not that it does. We can't just test that no events are registered, because
that might happen for a number of reasons (ie the underlying watcher service
failed). So instead we have a canary event alongside every operation that
should not generate an event, and then we check that the only event that is
registered is the canary event.

Signed-off-by: Moses Mendoza <moses@puppet.com>