(TK-448) make recursive an attribute of the watcher
For the watcher protocol implementation, add a new field `recursive`. The field
is initially populated as an atom with a nil value. Upon the first watch
directory added, the field is set "permanently" to the value specified for the
:recursive option. We do so by enforcing that upon addition of any new
directories to watch, the :recursive option must be the same as previously
supplied, or we throw an exception. Also add a docstring to
watch-new-directories! to clarify its behavior.

Signed-off-by: Moses Mendoza <moses@puppet.com>