(TK-448) Plumb recursive option through service
This commit updates the required methods in the watch service to take the
recursive setting into acount when deciding whether or not to recurse into
directories that are watched and add them to the list of watched directories.
With this, if a user specificies :recursive false, we limit watching to only
the directory specified by `add-watch-dir!` and no others.

This commit rounds out additional tests for this functionality.

Signed-off-by: Moses Mendoza <moses@puppet.com>