(PE-27497) Continue traversing directory when a file disappears
This commit updates our FileVisitor to continue iterating when it
encounters a file that has disappeared. If we have a watcher on a
directory with files managed by Puppet, there's a chance that we
might iterate over it while the file is being replaced, which
creates and then destroys a lockfile. Previously, this would cause the
Visitor to throw an exception, so this commit squashes that exception.