(maint) update clj-parent to latest, avoid expensive logging
This updates clj-parent to 4.20 and removes an exclusion for clojure.

It also conditionalizes some potentially expensive logging when
the logging levels aren't used. Specifically it avoids a `doseq` and
`count` of the events (which can be thousands as seen in the field) when
debug isn't enabled, and avoids a potentially expensive `print-to-string`
when trace isn't enabled.