(PE-37592) Compress events for the same changed-path
Previously, we would pass all events into a registered callback which
triggered for the given watch-path. However, in some cases we may
receive a `:create` event and then a `:modify` event as a new file is
written to disk. We could also receive multiple `:modify` events, and
because all of the events received from a WatchKey can be in any order
we may receive the `:create` event after a `:modify` event. Eg,
[`:modify`, `:create`, `:modify`, `:modify`] (imagine that these vectors
are Events all sharing the same changed-path).

However, we want to compress modifies into any creation events. This
will reduce the events given to callbacks. For example, the events
[`:create`, `:modify`] will reduce to only [`:create`], while
[`:modify`, `:create`] will also reduce to [`:create`]; events like
[`:modify`, `:modify`, `:modify`] will reduce to [`:modify`]. We do not
compress delete events, so events like [`:modify`, `:create`, `:modify`,
`:delete`] will be reduced to [`:create`, `:delete`].

The tests had the pecular property of collecting events until a given
event was received, then only returning _the given event_. The tests
have been updated to now return all the events received. Which, without
this patch, many tests fail because they receive modify and create
events, when the test is only expecting a create event. The tests now
pass with few changes because of the implemented event compression.